[BE] feat: WebSocket DISCONNECT시 토큰 못받던 문제 명시해서 해결 (#22)
